+++ frontend/src/logos/moon.png
